Projekt02/requirements.txt
flask
flask-sqlalchemy
flask-login
requests
werkzeug
pytest